Fix in scanner and reader

+++ src/NumberScanner.h
@@ -54,7 +54,7 @@ private:
     ucs4char* token_;
     ucs4char* limit_;
     ucs4char* marker_;
-    int bufferSize_{32};
+    size_t bufferSize_{32};
     enum conditon {
         INITIAL,
         IN_HEX,
